+++ RELEASE_NOTES.md
@@ -17,6 +17,7 @@
 - Design edit alarm page (#175)
 - Design alarm widget (#176)
 - Alarm page translations (#180)
+- JSON Mockup implementation (#182)
 
 ### Bug fixes
 
@@ -24,6 +25,7 @@
 - Datepicker fix for MUIx 6.0.2
 - Tenant Form Fix (new hook behavior)
 - react-colorful 5.6.1 (hook behavior)
+- not necessary state on sort button
 
 ### Documentation